Simplify custom build and task metrics
Remove the map fields that store the values for custom build and task
metrics. Instead, add them directly into metrics as SimpleMetrics.

Note that this introduces a behaviour change. Previously, event if you
set both build or task metrics, the defaults would be added.
This behaviour goes against the comment in the source that states "If
there was any customization then we assume that user populated
everything manually".

These changes will make that statement true, meaning that if either
build or task metrics are added, the default won't be and the user must
add those themselves.